correct get client requirements (#2)
A review of our current models showed that the `get_client` method of the `AuthProvider` model is currently required to implement in an Auth Provider implementation subclass while the documentation said that it was optional 

This was an oversight in the codding and needed to be adjusted. `get_client` is an optional method. however it was identified it is necessary to the use of clients issueing their own tokens when not going through an authentication cycle (using an existing toke to issue a token).

The fix was to change it in the code that `get_client` will be optional. However, if the method is not implemented the route enabling a regular user to issue themselves a token with a token will be disabled and raise a `409` if the active auth provider does not implement `get_client`

##Branch Commits:

* 1 removed client_data implementation requirement. Added tests for get_client implementation in default auth provider

* #1 corrected get_client method in builtin auth provider

* #1 added test for updated create_token response for non admin clients when get_client is not implemented, and updated route

* #1 updating version in setup.py

* #1 refactored test to be more explicit

* #1 updated langauge around get_client in docs to be more explicit about what is is used for